Splitting out non-core functionality from the server, removing redundant tests.
