+++ callouts/python/extproc/tests/add_header_test.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from __future__ import print_function
-
 import datetime
 from http.server import BaseHTTPRequestHandler
 from http.server import HTTPServer
@@ -20,13 +19,14 @@ import threading
 import time
 import urllib.request
 
+from envoy.service.ext_proc.v3 import external_processor_pb2 as service_pb2
+from envoy.service.ext_proc.v3 import external_processor_pb2_grpc as service_pb2_grpc
 import grpc
 from grpc import ServicerContext
 import pytest
 
 from extproc.service import callout_server
-from envoy.service.ext_proc.v3 import external_processor_pb2 as service_pb2
-from envoy.service.ext_proc.v3 import external_processor_pb2_grpc as  service_pb2_grpc
+from extproc.service import callout_tools
 
 # Global server variable.
 server: callout_server.CalloutServer | None = None
@@ -38,7 +38,7 @@ class CalloutServerTest(callout_server.CalloutServer):
       self, headers: service_pb2.HttpHeaders, context: ServicerContext
   ) -> service_pb2.HeadersResponse:
     """Custom processor on request headers."""
-    return callout_server.add_header_mutation(
+    return callout_tools.add_header_mutation(
       add=[('header-request', 'request')],
       clear_route_cache=True
     )
@@ -47,7 +47,7 @@ class CalloutServerTest(callout_server.CalloutServer):
       self, headers: service_pb2.HttpHeaders, context: ServicerContext
   ) -> service_pb2.HeadersResponse:
     """Custom processor on response headers."""
-    return callout_server.add_header_mutation(
+    return callout_tools.add_header_mutation(
       add=[('header-response', 'response')],
       remove=['foo']
     )
@@ -107,14 +107,14 @@ class TestBasicServer(object):
 
         value = _MakeRequest(stub, response_headers=headers, async_mode=False)
         assert value.HasField('response_headers')
-        assert value.response_headers == callout_server.add_header_mutation(
+        assert value.response_headers == callout_tools.add_header_mutation(
           add=[('header-response', 'response')],
           remove=['foo']
         )
 
         value = _MakeRequest(stub, request_headers=headers, async_mode=False)
         assert value.HasField('request_headers')
-        assert value.request_headers == callout_server.add_header_mutation(
+        assert value.request_headers == callout_tools.add_header_mutation(
           add=[('header-request', 'request')],
           clear_route_cache=True
         )
